agregue un fragment para ver los trailer, y puse una funcionalidad para que un usuario no pueda comprar 2 veces la misma pelicula

+++ app/src/main/java/com/example/tp_appsmoviles_grupof/database/local/entities/userEntity.kt
@@ -4,12 +4,14 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 import androidx.room.ColumnInfo
 
+
 @Entity(tableName = "users")
 data class userEntity(
-    @PrimaryKey(autoGenerate = true) val idUser: Int = 0,
+    @PrimaryKey(autoGenerate = true)  val idUser: Long = 0L,
     @ColumnInfo(name = "nombre")   val nombre: String,
     @ColumnInfo(name = "password") val password: String,
     @ColumnInfo(name = "email")    val email: String? = null,
     @ColumnInfo(name = "telefono") val telefono: String? = null,
     @ColumnInfo(name = "direccion")val direccion: String? = null,
 )
+
